scripts/requirements.txt
# ==================================================
# Compliance Bot - Wykrywacz Danych Wrażliwych
# ==================================================
# Requirements for PII Detection and GDPR Compliance
# Author: Kasia Majzel-Pośpiech
# Date: 2025-11-10
# ==================================================

# Azure AI Services
azure-ai-contentsafety>=1.0.0
azure-ai-ml>=1.11.0
azure-identity>=1.15.0

# OpenAI
openai>=1.12.0

# Web Framework & API
flask>=3.0.0
requests>=2.31.0
werkzeug>=3.0.1

# Data Processing & Analysis
pandas>=2.1.0
openpyxl>=3.1.0

# Configuration & Utilities
python-dotenv>=1.0.0
colorama>=0.4.6

# ==================================================
# Installation:
# pip install -r requirements.txt
# ==================================================